\name{alice}
\alias{alice}
\docType{data}
\title{Alice's Adventures in Wonderland}
\description{
The text of Lewis Carroll's 'Alice's Adventures in Wonderland', with 
punctuation marks removed.
}
\usage{data(alice)}
\format{
  A character vector with 25942 words.
}
\source{
  The project Gutenberg at \url{http://www.gutenberg.org/wiki/Main_Page}
}
\examples{
  data(alice)
  alice[1:4]
}
\keyword{datasets}
