% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiInfo}
\alias{apiInfo}
\title{List API information for each function.}
\usage{
apiInfo(funName)
}
\arguments{
\item{funName}{Function name as a string.}
}
\value{
Print the information of API keys for functions in the package.
}
\examples{
apiInfo()
}
\references{
See \url{https://azure.microsoft.com/en-us/services/cognitive-services/}.
}

