\name{corres-class}
\docType{class}
\alias{corres-class}

\title{Class "corres" }
\description{A class for correspondence analysis}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("corres", ...)}.
	Correspondence objects can be plotted, summarized, and printed. 
}
\section{Slots}{
	 \describe{
    \item{\code{data}:}{Object of class \code{"list"} ~~ }
  }
}
\section{Methods}{
No methods defined with class "corres" in the signature.
}
\references{ Murtagh }
\author{R. H. Baayen}
\note{to be expanded}

\seealso{
	See Also \code{corres.fnc}.
}
\examples{
showClass("corres")
}
\keyword{classes}
