\name{corresInit}
\alias{corresInit}
\title{Initialize correspondence object}
\description{
  Function initializing correspondence object.
}
\usage{
corresInit(data)
}
\arguments{
  \item{data}{A list, see the code of \code{corres} for further details.}
}
\value{
  A correspondence object.
}
\author{R. Harald Baayen}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
