\name{getPos.fnc}
\alias{getPos.fnc}
\title{ determine position for labels for interaction plots }
\description{
  determines the position (in the X and Y vectors) for the 
  adding of text to an interaction plot 
}
\usage{
getPos.fnc(vec, pos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{vector of Y values}
  \item{pos}{can be '"beg"', '"mid"', '"end"'}
}
\details{
  not intended for independent use
}
\value{
  an integer specifying position in vector for X and Y values in plot
}
\references{ ~put references to the literature/web site here ~ }
\author{R. H. Baayen}
\note{
not indended for independent use
}
\seealso{See Also as \code{\link{plotLMER.fnc}}}
\examples{
  \dontrun{not intended for independent use}
}
\keyword{ regression }
