\name{getRange.fnc}
\alias{getRange.fnc}
\title{Extracts range of predicted values from list of data frames}
\description{
  Extracts range of predicted values from list of data frames
}
\usage{
getRange.fnc(lst)
}
\arguments{
  \item{lst}{a list with one or more data frames with column names
  \code{Y} and optionally \code{lower} and \code{upper}.}
}
\details{
  not intended for independent use.
}
\value{
  \item{value}{a two-element vector specifying the range of values in \code{Y}}
}
\references{ ~put references to the literature/web site here ~ }
\author{R. H. Baayen}
\note{
not intended for separate use.
}
\seealso{ See Also as \code{\link{plotLMER.fnc}}}
\examples{
\dontrun{
  not intended for independent use
}
}
\keyword{ regression }
