\name{getRoot.fnc}
\alias{getRoot.fnc}
\title{extract simple name of predictor from expression with poly}
\description{
  extract \code{X} from expressions such as \code{poly(X, 3, raw = TRUE}
}
\usage{
getRoot.fnc(xlabel)
}
\arguments{
  \item{xlabel}{character string for predictor name}
}
\details{
 not intended for independent use
}
\value{
  a character string (simple name of predictor)
}
%\references{ ~put references to the literature/web site here ~ }
\author{R. H. Baayen}
\note{ 
  not intended for independent use
}
\seealso{See Also as \code{\link{plotLMER.fnc}}}
\examples{
  \dontrun{not intended for independent use}
}
\keyword{ regression }
