\name{growth-class}
\docType{growth}
\alias{growth-class}

\title{Class "growth" }
\description{A class for the analysis of word frequency distributions}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("growth", ...)}.
	Growth objects can be plotted, summarized, and printed. 
}
\section{Slots}{
	 \describe{
    \item{\code{data}:}{Object of class \code{"list"}  }
  }
}
\section{Methods}{
No methods defined with class "growth" in the signature.
}
\references{ R. H. Baayen, 2007 }
\author{R. H. Baayen}
\note{to be expanded}

\seealso{
	See Also \code{growth.fnc}.
}
\examples{
showClass("corres")
}
\keyword{classes}
