\name{growthInit}
\alias{growthInit}
\title{Initialize a vocabulary growth object.}
\description{
  This function initializes a growth object.
}
\usage{
growthInit(data)
}
\arguments{
  \item{data}{A list with as single element a data frame with columns as
        described in the documentation of growth.fnc.}
}
\value{
  A growth object, see growth.fnc for details.
}
\author{R. H. Baayen}
\seealso{See Also as \code{\link{growth.fnc}}.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
