\name{implementInteractions.fnc}
\alias{implementInteractions.fnc}
\title{implement interactions in the model matrix}
\description{
  given a model matrix with main effects only, add interactions
}
\usage{
implementInteractions.fnc(m)
}
\arguments{
  \item{m}{a (model) matrix (rows observations, columns predictors)}
}
\details{
  not intended for independent use
}
\value{
  an updated (model) matrix
}
%\references{ ~put references to the literature/web site here ~ }
\author{R. H. Baayen}
\note{not intended for independent use}
\seealso{\code{\link{plotLMER.fnc}}}
\examples{
  \dontrun{not intended for independent use}
}
\keyword{ regression }
