\name{moby}
\alias{moby}
\docType{data}
\title{Moby Dick}
\description{
The text of H. Melville's 'Moby Dick', with 
punctuation marks removed.
}
\usage{data(alice)}
\format{
  A character vector with 215994 words.
}
\source{
  The project Gutenberg at \url{http://www.gutenberg.org/wiki/Main_Page}
}
\examples{\dontrun{
  data(moby)
  moby[1:2]
}}
\keyword{datasets}
