% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{outputFromResponse}
\alias{outputFromResponse}
\title{Manipulate response to data frame.}
\usage{
outputFromResponse(response)
}
\arguments{
\item{r}{response returned from a http request.}
}
\value{
response in a friendly format (e.g., data frame or list).
}

