\name{oz}
\alias{oz}
\docType{data}
\title{The Wonderful Wizard of Oz}
\description{
The text of L. F. Baum's 'The Wonderful Wizard of Oz', with 
punctuation marks removed.
}
\usage{data(alice)}
\format{
  A character vector with 38177 words.
}
\source{
  The project Gutenberg at \url{http://www.gutenberg.org/wiki/Main_Page}
}
\examples{\dontrun{
  data(oz)
  oz[1:5]
}}
\keyword{datasets}
