\name{transforming.fnc}
\alias{transforming.fnc}
\title{transform vector according to specified function}
\description{
  Apply function fun to input vector y
}
\usage{
transforming.fnc(y, fun)
}
\arguments{
  \item{y}{ numerical vector (for dependent variable)}
  \item{fun}{a function, or NA (in which case no transformation is applied}
}
\details{
  exists only to make code more readable
}
\value{
  a numerical vector
}
\author{R. H. Baayen}
\note{ 
 not intended for independent use
}
\seealso{See Also as \code{\link{plotLMER.fnc}}}
\examples{
  \dontrun{not intended for independent use}
}
  \keyword{ regression }
