\name{through}
\alias{through}
\docType{data}
\title{Through the Looking Glass}
\description{
The text of Lewis Carroll's 'Through the Looking Glass', with 
punctuation marks removed.
}
\usage{data(through)}
\format{
  A character vector with 29560 words.
}
\source{
  The project Gutenberg at \url{http://www.gutenberg.org/wiki/Main_Page}
}
\examples{\dontrun{
  data(through)
  through[1:3]
}}
\keyword{datasets}
